requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.128.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0

Faker==38.2.0
mysql-connector-python==9.5.0
numpy==2.3.5
pandas==2.3.3
py4j==0.10.9.9
pyspark==4.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2

